+++ src/main/java/com/acn/users/service/UserServiceImpl.java
@@ -4,6 +4,7 @@ import com.acn.users.model.User;
 import com.acn.users.repository.UserRepository;
 import org.springframework.stereotype.Service;
 
+import javax.persistence.EntityNotFoundException;
 import java.util.NoSuchElementException;
 import java.util.Optional;
 
@@ -18,10 +19,14 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public User getUserById(Long id) throws NoSuchElementException {
+
         Optional<User> retrievedUser = userRepository.findById(id);
 
-        if(retrievedUser.isPresent())   return retrievedUser.get();
-        else                            throw new NoSuchElementException();
+        if(retrievedUser.isPresent()) {
+            return retrievedUser.get();
+        } else {
+            throw new NoSuchElementException("Could not find item with id " + id + ". Please verify that your parameters are set correctly.");
+        }
     }
 
     @Override
@@ -31,26 +36,36 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public User updateUser(Long id, User user) {
-        User    userToUpdate    = userRepository.getOne(id);
-        String  userFirstName   = user.getFirstname();
-        String  userLastName    = user.getLastname();
-        String  userAddress     = user.getAddress();
-        String  userEmail       = user.getEmail();
-        String  userPhone       = user.getPhone();
+        if(userRepository.existsById(id)) {
+            User    userToUpdate    = userRepository.getOne(id);
+            String  userFirstName   = user.getFirstname();
+            String  userLastName    = user.getLastname();
+            String  userAddress     = user.getAddress();
+            String  userEmail       = user.getEmail();
+            String  userPhone       = user.getPhone();
 
-        // update the required fields
-        if(userFirstName != null) userToUpdate.setFirstname(userFirstName);
-        if(userLastName  != null) userToUpdate.setLastname(userLastName);
-        if(userAddress   != null) userToUpdate.setAddress(userAddress);
-        if(userEmail     != null) userToUpdate.setEmail(userEmail);
-        if(userPhone     != null) userToUpdate.setPhone(userPhone);
+            // update the required fields
+            if(userFirstName != null) userToUpdate.setFirstname(userFirstName);
+            if(userLastName  != null) userToUpdate.setLastname(userLastName);
+            if(userAddress   != null) userToUpdate.setAddress(userAddress);
+            if(userEmail     != null) userToUpdate.setEmail(userEmail);
+            if(userPhone     != null) userToUpdate.setPhone(userPhone);
 
-        // save the user
-        return userRepository.save(userToUpdate);
+            // save the user
+            return userRepository.save(userToUpdate);
+
+        } else {
+            throw new EntityNotFoundException("Could not update user. User with id " + id + " does not exist");
+        }
     }
 
     @Override
     public void deleteUserById(Long id) {
-        userRepository.deleteById(id);
+        if(userRepository.existsById(id)) {
+            userRepository.deleteById(id);
+        } else {
+            throw new EntityNotFoundException("Could not perform delete operation. User with id " + id + " could not be found or doesn't exist.");
+        }
+
     }
 }
